+++ eq-author/src/App/QuestionnaireDesignPage/NavigationSidebar/index.js
@@ -6,13 +6,8 @@ import CustomPropTypes from "custom-prop-types";
 import styled from "styled-components";
 import { colors } from "constants/theme";
 
-import {
-  buildSectionPath,
-  buildFolderPath,
-  buildPagePath,
-  buildConfirmationPath,
-  buildIntroductionPath,
-} from "utils/UrlUtils";
+import { buildIntroductionPath } from "utils/UrlUtils";
+
 import {
   getPageById,
   getFolderById,
@@ -20,18 +15,14 @@ import {
   getSectionByFolderId,
 } from "utils/questionnaireUtils";
 
-import { DragDropContext, Droppable, Draggable } from "react-beautiful-dnd";
+import { DragDropContext } from "react-beautiful-dnd";
 import NavigationHeader from "./NavigationHeader";
-import CollapsibleNavItem from "components/CollapsibleNavItem";
 import NavItem from "components/NavItem";
 import ScrollPane from "components/ScrollPane";
 import Button from "components/buttons/Button";
 
-import IconSection from "assets/icon-section.svg?inline";
-import IconFolder from "assets/icon-folder.svg?inline";
-import IconQuestionPage from "assets/icon-questionpage.svg?inline";
-import IconConfirmationPage from "assets/icon-playback.svg?inline";
-import IconSummaryPage from "assets/icon-summarypage.svg?inline";
+import Section from "./Section";
+
 import PageIcon from "assets/icon-survey-intro.svg?inline";
 
 import MOVE_PAGE_MUTATION from "graphql/movePage.graphql";
@@ -87,197 +78,6 @@ const IntroductionListItem = styled(ListItem)`
     font-weight: bold;
   }
 `;
-
-const Page = ({
-  id: pageId,
-  questionnaireId,
-  displayName,
-  pageType,
-  confirmation,
-  validationErrorInfo,
-  position,
-}) => {
-  const { entityId, tab = "design" } = useParams();
-
-  const isCurrentPage = (navItemId, currentPageId) =>
-    navItemId === currentPageId;
-
-  const iconMap = {
-    QuestionPage: IconQuestionPage,
-    CalculatedSummaryPage: IconSummaryPage,
-  };
-
-  return (
-    <Draggable key={pageId} draggableId={pageId} index={position}>
-      {({ innerRef, draggableProps, dragHandleProps }, { isDragging }) => (
-        <ListItem ref={innerRef} {...draggableProps}>
-          <NavItem
-            title={displayName}
-            icon={iconMap[pageType]}
-            disabled={isCurrentPage(pageId, entityId)}
-            titleUrl={buildPagePath({
-              questionnaireId,
-              pageId,
-              tab,
-            })}
-            errorCount={validationErrorInfo?.totalCount}
-            isDragging={isDragging}
-            {...dragHandleProps}
-          />
-          {confirmation && (
-            <NavItem
-              key={confirmation.displayName}
-              title={confirmation.displayName}
-              titleUrl={buildConfirmationPath({
-                questionnaireId,
-                confirmationId: confirmation.id,
-                tab,
-              })}
-              disabled={isCurrentPage(confirmation.id, entityId)}
-              icon={IconConfirmationPage}
-              errorCount={confirmation?.validationErrorInfo?.totalCount}
-              {...dragHandleProps}
-            />
-          )}
-        </ListItem>
-      )}
-    </Draggable>
-  );
-};
-
-const Folder = ({
-  id: folderId,
-  questionnaireId,
-  displayName,
-  pages,
-  validationErrorInfo,
-  position,
-}) => {
-  const { entityId, tab = "design" } = useParams();
-
-  const isCurrentFolder = (navItemId, currentPageId) =>
-    navItemId === currentPageId;
-
-  const calculatePageErrors = (pages) =>
-    pages.reduce(
-      (acc, { validationErrorInfo }) => (acc += validationErrorInfo.totalCount),
-      0
-    );
-
-  return (
-    <Draggable key={folderId} draggableId={folderId} index={position}>
-      {({ innerRef, draggableProps, dragHandleProps }, { isDragging }) => (
-        <ListItem ref={innerRef} {...draggableProps} {...dragHandleProps}>
-          <CollapsibleNavItem
-            title={displayName}
-            icon={IconFolder}
-            disabled={isCurrentFolder(folderId, entityId)}
-            titleUrl={buildFolderPath({
-              questionnaireId,
-              folderId,
-              tab,
-            })}
-            selfErrorCount={validationErrorInfo.totalCount}
-            childErrorCount={calculatePageErrors(pages)}
-            isDragging={isDragging}
-          >
-            <Droppable droppableId={folderId} type="folderContent">
-              {({ innerRef, placeholder, droppableProps }) => (
-                <NavList ref={innerRef} {...droppableProps}>
-                  {pages.map(({ id: pageId, ...rest }) => (
-                    <Page
-                      key={`page-${pageId}`}
-                      id={pageId}
-                      questionnaireId={questionnaireId}
-                      {...rest}
-                    />
-                  ))}
-                  {placeholder}
-                </NavList>
-              )}
-            </Droppable>
-          </CollapsibleNavItem>
-        </ListItem>
-      )}
-    </Draggable>
-  );
-};
-
-const Section = ({
-  id: sectionId,
-  questionnaireId,
-  displayName,
-  folders,
-  validationErrorInfo,
-}) => {
-  const { entityId, tab = "design" } = useParams();
-
-  const isCurrentPage = (navItemId, currentPageId) =>
-    navItemId === currentPageId;
-
-  const calculatePageErrors = (pages) =>
-    pages.reduce(
-      (acc, { validationErrorInfo }) => (acc += validationErrorInfo.totalCount),
-      0
-    );
-
-  const allPagesInSection = folders.flatMap(({ pages }) => pages);
-
-  const renderChild = ({ id: folderId, enabled, pages, position, ...rest }) => {
-    if (enabled) {
-      return (
-        <Folder
-          key={`folder-${folderId}`}
-          id={folderId}
-          questionnaireId={questionnaireId}
-          pages={pages}
-          position={position}
-          {...rest}
-        />
-      );
-    } else {
-      return pages.map(({ id: pageId, ...rest }) => (
-        <Page
-          key={`page-${pageId}`}
-          id={pageId}
-          questionnaireId={questionnaireId}
-          {...rest}
-          position={position}
-        />
-      ));
-    }
-  };
-
-  return (
-    <ListItem>
-      <CollapsibleNavItem
-        title={displayName}
-        icon={IconSection}
-        bordered
-        disabled={isCurrentPage(sectionId, entityId)}
-        titleUrl={buildSectionPath({
-          questionnaireId,
-          sectionId,
-          tab,
-        })}
-        selfErrorCount={validationErrorInfo.totalCount}
-        childErrorCount={calculatePageErrors(allPagesInSection)}
-        containsActiveEntity={allPagesInSection
-          .map(({ id }) => isCurrentPage(id, entityId))
-          .find(Boolean)}
-      >
-        <Droppable droppableId={sectionId} type="sectionContent">
-          {({ innerRef, placeholder, droppableProps }) => (
-            <NavList ref={innerRef} {...droppableProps}>
-              {folders.map((folder) => renderChild(folder))}
-              {placeholder}
-            </NavList>
-          )}
-        </Droppable>
-      </CollapsibleNavItem>
-    </ListItem>
-  );
-};
 
 const NavigationSidebar = ({ questionnaire }) => {
   const { entityId, tab = "design" } = useParams();
